pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deadpgp"
version = "0.1.0"
description = "Policy-gated GPG reveal workflow with audit logging"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
deadpgp-reveal = "deadpgp.workflow:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["deadpgp*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
